feat: introduce release channels
vergen is rather heavy, and the ugly `0.1.0 (VERGEN_IDEMPOTENT_OUTPUT)`
placeholder in the About dialog looks dumb. instead, we introduce the
concept of "release channels" which determine how to format the version
string. currently, the accepted variants are: "dev" "stable" "flake". can
be set via build-time env vars or the `package/RELEASE_CHANNEL` file.

this adds a build-dependency on `git` cli, though that's arguably more
acceptable than building all of `libgit2` as before. however, the commit
hash is no longer shown in stable releases, so this is really only a
dependency for developers / pre-release users (i.e., when knowing the
exact commit id would actually matter). for these users, hummingbird will
fail to build if git isn't installed or the `.git/` directory isn't
present.

to make it necessary to run the following before publishing a release:
`echo "stable" > hummingbird/package/RELEASE_CHANNEL`
1. go to settings > rules > rulesets > new ruleset > new tag ruleset
2. set name, enforcement status, target tags etc
3. select "require status checks to pass" > add checks
4. type `validate-release-channel` and set the source to github actions

with this, releases cannot be created unless `check-release-channel.yml`
succeeds. the workflow is automatically triggered on PRs/pushes targeting
a branch named `release`, which can serve as a prep area for staging new
releases. this can of course be changed, this is just the simplest method
i could think of that wouldn't make `master`'s ci permanently red for no
reason.